pkg/scaffold: move diff inside scaffold and unexport it
