+++ pkg/scaffold/addtoscheme_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestAddToScheme(t *testing.T) {
@@ -32,7 +30,7 @@ func TestAddToScheme(t *testing.T) {
 	}
 
 	if addtoschemeExp != buf.String() {
-		diffs := util.Diff(addtoschemeExp, buf.String())
+		diffs := diff(addtoschemeExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
