+++ pkg/scaffold/apis_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestApis(t *testing.T) {
@@ -28,7 +26,7 @@ func TestApis(t *testing.T) {
 	}
 
 	if apisExp != buf.String() {
-		diffs := util.Diff(apisExp, buf.String())
+		diffs := diff(apisExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
