+++ pkg/scaffold/build_dockerfile_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestDockerfile(t *testing.T) {
@@ -28,7 +26,7 @@ func TestDockerfile(t *testing.T) {
 	}
 
 	if dockerfileExp != buf.String() {
-		diffs := util.Diff(dockerfileExp, buf.String())
+		diffs := diff(dockerfileExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
