+++ pkg/scaffold/gitignore_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestGitignore(t *testing.T) {
@@ -28,7 +26,7 @@ func TestGitignore(t *testing.T) {
 	}
 
 	if gitignoreExp != buf.String() {
-		diffs := util.Diff(gitignoreExp, buf.String())
+		diffs := diff(gitignoreExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
