+++ pkg/scaffold/operator_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestOperator(t *testing.T) {
@@ -28,7 +26,7 @@ func TestOperator(t *testing.T) {
 	}
 
 	if operatorExp != buf.String() {
-		diffs := util.Diff(operatorExp, buf.String())
+		diffs := diff(operatorExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
