+++ pkg/scaffold/register_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestRegister(t *testing.T) {
@@ -32,7 +30,7 @@ func TestRegister(t *testing.T) {
 	}
 
 	if registerExp != buf.String() {
-		diffs := util.Diff(registerExp, buf.String())
+		diffs := diff(registerExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
