+++ pkg/scaffold/test_framework_dockerfile_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestTestFrameworkDockerfile(t *testing.T) {
@@ -28,7 +26,7 @@ func TestTestFrameworkDockerfile(t *testing.T) {
 	}
 
 	if testFrameworkDockerfileExp != buf.String() {
-		diffs := util.Diff(testFrameworkDockerfileExp, buf.String())
+		diffs := diff(testFrameworkDockerfileExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
