+++ pkg/scaffold/version_test.go
@@ -16,8 +16,6 @@ package scaffold
 
 import (
 	"testing"
-
-	"github.com/operator-framework/operator-sdk/pkg/util"
 )
 
 func TestVersion(t *testing.T) {
@@ -28,7 +26,7 @@ func TestVersion(t *testing.T) {
 	}
 
 	if versionExp != buf.String() {
-		diffs := util.Diff(versionExp, buf.String())
+		diffs := diff(versionExp, buf.String())
 		t.Fatalf("expected vs actual differs.\n%v", diffs)
 	}
 }
